feat: implement image scrolling and pagination
#41
